requirements.txt
aiogram==3.20.0
firebase-admin==6.5.0
python-dotenv==1.0.1
